requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
defusedxml==0.7.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
et_xmlfile==2.0.0
greenlet==3.3.2
idna==3.11
oauthlib==3.3.1
openpyxl==3.1.5
pillow==12.1.1
pycparser==3.0
pyee==13.0.1
PyJWT==2.12.1
python3-openid==3.2.0
reportlab==4.4.10
requests==2.33.0
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
